// types/enums.ts
export enum SortField {
  Position = 'position',
  Name = 'name',
  Points = 'points',
  Wins = 'wins',
  Draws = 'draws',
  Losses = 'losses',
  GoalsFor = 'goalsFor',
  GoalsAgainst = 'goalsAgainst',
}

export enum SortDirection {
  Asc = 'asc',
  Desc = 'desc',
}

export enum MatchResult {
  Win = 'W',
  Draw = 'D',
  Loss = 'L',
}